+++ dio/lib/src/utils.dart
@@ -21,8 +21,8 @@ const kReleaseMode = bool.fromEnvironment('dart.vm.product');
 /// Pipes all data and errors from [stream] into [sink]. Completes [Future] once
 /// [stream] is done. Unlike [store], [sink] remains open after [stream] is
 /// done.
-Future writeStreamToSink(Stream stream, EventSink sink) {
-  final completer = Completer();
+Future<void> writeStreamToSink<T>(Stream<T> stream, EventSink<T> sink) {
+  final completer = Completer<void>();
   stream.listen(
     sink.add,
     onError: sink.addError,