+++ dio/lib/dio.dart
@@ -20,4 +20,5 @@ export 'src/redirect_record.dart';
 export 'src/response.dart';
 export 'src/transformer.dart';
 export 'src/transformers/background_transformer.dart';
+export 'src/transformers/fused_transformer.dart';
 export 'src/transformers/sync_transformer.dart';