+++ dio/lib/src/transformer.dart
@@ -5,6 +5,12 @@ import 'adapter.dart';
 import 'options.dart';
 import 'utils.dart';
 
+/// The callback definition for decoding a JSON string.
+typedef JsonDecodeCallback = FutureOr<dynamic> Function(String);
+
+/// The callback definition for encoding a JSON object.
+typedef JsonEncodeCallback = FutureOr<String> Function(Object);
+
 /// [Transformer] allows changes to the request/response data before
 /// it is sent/received to/from the server.
 ///
@@ -72,5 +78,29 @@ abstract class Transformer {
     return mediaType.mimeType == 'application/json' ||
         mediaType.mimeType == 'text/json' ||
         mediaType.subtype.endsWith('+json');
+  }
+
+  static FutureOr<String> defaultTransformRequest(
+    RequestOptions options,
+    JsonEncodeCallback jsonEncodeCallback,
+  ) {
+    final Object data = options.data ?? '';
+    if (data is! String && Transformer.isJsonMimeType(options.contentType)) {
+      return jsonEncodeCallback(data);
+    } else if (data is Map) {
+      if (data is Map<String, dynamic>) {
+        return Transformer.urlEncodeMap(data, options.listFormat);
+      }
+      debugLog(
+        'The data is a type of `Map` (${data.runtimeType}), '
+        'but the transformer can only encode `Map<String, dynamic>`.\n'
+        'If you are writing maps using `{}`, '
+        'consider writing `<String, dynamic>{}`.',
+        StackTrace.current,
+      );
+      return data.toString();
+    } else {
+      return data.toString();
+    }
   }
 }