+++ dio/CHANGELOG.md
@@ -8,6 +8,7 @@ See the [Migration Guide][] for the complete breaking changes list.**
 - Raise the min Dart SDK version to 2.18.0.
 - Add constructor for `DioExceptionType.badCertificate`.
 - Create type alias `DioMediaType` for `http_parser`'s `MediaType`.
+- Fix the type conversion regression when using `MultipartFile.fromBytes`.
 
 ## 5.4.3+1