+++ dio/CHANGELOG.md
@@ -17,6 +17,7 @@ See the [Migration Guide][] for the complete breaking changes list.**
   throws the exception directly rather than cut actual HTTP requests afterward.
 - Catch `MediaType` parse exception in `Transformer.isJsonMimeType`.
 - Improves warning logs on the Web platform.
+- Improves memory allocating when using `CancelToken`.
 
 ## 5.4.3+1